refactor: update pantry aggregation imports
