+++ MJ_FB_Backend/tests/clientVisitController.test.ts
@@ -4,7 +4,7 @@ import {
   refreshPantryWeekly,
   refreshPantryMonthly,
   refreshPantryYearly,
-} from '../src/controllers/pantryAggregationController';
+} from '../src/controllers/pantry/pantryAggregationController';
 import {
   addVisit,
   deleteVisit,
@@ -12,7 +12,7 @@ import {
   getVisitStats,
 } from '../src/controllers/clientVisitController';
 
-jest.mock('../src/controllers/pantryAggregationController', () => ({
+jest.mock('../src/controllers/pantry/pantryAggregationController', () => ({
   refreshPantryWeekly: jest.fn(),
   refreshPantryMonthly: jest.fn(),
   refreshPantryYearly: jest.fn(),
